+++ src/models/deviceSettings/index.ts
@@ -3,6 +3,7 @@ export * from './actuatorSettings';
 export * from './cameraSettings';
 export * from './deviceSettings';
 export * from './dimmerSettings';
+export * from './garageDoorOpenerSettings';
 export * from './heaterSettings';
 export * from './ledSettings';
 export * from './motionSensorSettings';