+++ src/server/devices/baseDeviceInterfaces/iGarageDoorOpener.ts
@@ -1,0 +1,14 @@
+import { GarageDoorOpenerSettings } from '../../../models';
+import { iRoomDevice } from './iRoomDevice';
+
+export interface iGarageDoorOpener extends iRoomDevice {
+  settings: GarageDoorOpenerSettings;
+
+  readonly isClosed: boolean;
+
+  open(): void;
+
+  close(): void;
+
+  trigger(): void;
+}