+++ src/server/devices/index.ts
@@ -8,6 +8,7 @@ export * from './jsObject/index';
 export * from './models/index';
 export * from './scene/index';
 export * from './shelly/index';
+export * from './tuya/index';
 export * from './zigbee/index';
 export * from './device-cluster';
 export * from './device-cluster-type';