+++ src/server/devices/tuya/index.ts
@@ -1,0 +1,2 @@
+export { TuyaDevice } from './tuyaDevice';
+export { TuyaGarageOpener } from './tuyaGarageOpener';